pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ibdnet"
version = "0.1.0"
description = "IBD detection using PBWT-derived features and neural CRF"
authors = [
  { name = "IBDNet Developers" }
]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
keywords = ["genomics", "ibd", "pbwt", "hmm", "machine-learning"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]

dependencies = [
  "numpy>=1.26",
  "numba>=0.59",
  "polars>=0.19",
  "tskit>=0.5",
  "msprime>=1.3",
  "torch>=2.2",
  "torch-struct>=0.5",
  "zarr>=2.16",
  "cyvcf2>=0.30",
  "pybedtools>=0.9",
  "pyranges>=0.0.129",
  "pandas>=2.1",
  "scipy>=1.11",
  "click>=8.1",
  "tqdm>=4.66",
  "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-cov>=4.1",
  "mypy>=1.8"
]

sim = [
  "stdpopsim>=0.3",
]

[tool.setuptools.packages.find]
include = ["ibdnet"]
where = ["."]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
